requirements.txt
crawl4ai
google-genai
pydantic
python-dotenv
pytest
pytest-asyncio
playwright
playwright-stealth
beautifulsoup4
discord.py